floorflour/requirements.txt
PySide6>=6.6.0
PyYAML>=6.0
chardet>=5.0
